fixed mjs issues, converted everything to cjs with .js extention files. clerk and mongo now work good.
